% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boston.R
\docType{data}
\name{precincts}
\alias{precincts}
\title{Election Precincts of the City of Boston}
\format{A \code{SpatialPolygonsDataFrame} with 253 features and 3 attributes:
\describe{
\item{\code{precinct}}{integer precinct number.}
\item{\code{ward}}{integer number of ward in which precinct resides.}
\item{\code{id}}{factor unique identifier for each precinct.  Combines
\code{ward} and \code{precinct}.}
}

See
 \url{https://www.cityofboston.gov/maps/pdfs/ward_and_precincts.pdf} for details.}
\usage{
precincts
}
\description{
Spatial boundaries of the 253 election precincts for Boston.
}
\keyword{datasets}

