% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boston.R
\docType{data}
\name{wards}
\alias{wards}
\title{Election Wards of the City of Boston}
\format{A \code{SpatialPolygonsDataFrame} with 22 features and 2 attributes:
\describe{
\item{\code{ward}}{integer ward number.}
\item{\code{n_precincts}}{integer number of police precincts in ward.}
}

See
 \url{https://www.cityofboston.gov/maps/pdfs/ward_and_precincts.pdf} for details.}
\usage{
wards
}
\description{
Spatial boundaries of the 22 election wards for Boston.
}
\keyword{datasets}

